[DOC] README: Replace `*` with `×`
The `*` in `3*8 KB` is interpreted by some Markdown parsers
(particularly any following the CommonMark[1] standard) as the start
of an emphasis section. (The GitHub Markdown renderer does not follow
this standard, which is perhaps why it was not noticed until now.)
Thus it needs to be escaped or replaced.

We replace it with Unicode Multipication Sign (`×`, U+00D7) since
that's exactly the meaning we want to express here.

[1]: https://spec.commonmark.org/0.28/#emphasis-and-strong-emphasis